requirements.txt
absl-py==2.1.0
array_record==0.5.1
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
Cartopy==0.23.0
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
cloudpickle==3.0.0
clu==0.0.12
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
cython==3.1.0
debugpy==1.8.2
decorator==5.1.1
dm-tree==0.1.8
etils==1.9.2
executing==2.0.1
fasteners==0.19
flatbuffers==24.3.25
flax==0.8.5
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gin-config==0.5.0
google-pasta==0.2.0
grain-nightly==0.0.7
grpcio==1.65.2
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
jax==0.4.31
jax-cuda12-pjrt==0.4.31
jax-cuda12-plugin==0.4.31
jaxlib==0.4.31
jaxtyping==0.2.33
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
more-itertools==10.3.0
msgpack==1.0.8
namex==0.0.8
nest-asyncio==1.6.0
numcodecs==0.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opt-einsum==3.3.0
optax==0.2.3
optree==0.12.1
orbax-checkpoint==0.5.23
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
properscoring==0.1
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1

python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
-e git+https://git@github.com/limamau/s2s-downscaling.git@e26c8228f023cd3df99227584462d3261f0f9e18#egg=s2s_downscaling
scipy==1.14.0
shapely==2.0.5
six==1.16.0
stack-data==0.6.3
swirl_dynamics @ git+https://github.com/limamau/swirl-dynamics.git@d5f9a052e418195d2eb73e6392ef1dd19bfefd65
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.64
termcolor==2.4.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xarray==2024.7.0
xarray-tensorstore==0.1.1
zarr==2.18.2
zipp==3.19.2
